Issue #4 Made the project Android Agnostic and updated build tools
