--- gsonpath-extension-android/src/main/java/gsonpath/extension/android/range/intrange/AndroidIntRangeGsonPathFieldValidator.java
@@ -1,12 +1,0 @@
-package gsonpath.extension.android.range.intrange;
-
-import com.google.auto.service.AutoService;
-import gsonpath.compiler.GsonPathExtension;
-
-/**
- * Note: AutoService only works when annotating a Java class.
- */
-@SuppressWarnings("All")
-@AutoService(GsonPathExtension.class)
-public class AndroidIntRangeGsonPathFieldValidator extends AndroidIntRangeGsonPathFieldValidatorImpl {
-}
