+++ gsonpath-extensions-compiler/src/main/java/gsonpath/extension/def/range/intrange/IntRangeGsonPathFieldValidator.java
@@ -1,0 +1,12 @@
+package gsonpath.extension.def.range.intrange;
+
+import com.google.auto.service.AutoService;
+import gsonpath.compiler.GsonPathExtension;
+
+/**
+ * Note: AutoService only works when annotating a Java class.
+ */
+@SuppressWarnings("All")
+@AutoService(GsonPathExtension.class)
+public class IntRangeGsonPathFieldValidator extends IntRangeGsonPathFieldValidatorImpl {
+}
