+++ integration-test/build.gradle
@@ -1,4 +1,6 @@
 apply plugin: 'java'
+apply plugin: 'kotlin'
+apply plugin: 'kotlin-kapt'
 
 targetCompatibility = JavaVersion.VERSION_1_7
 sourceCompatibility = JavaVersion.VERSION_1_7
@@ -10,18 +12,18 @@ buildscript {
         }
     }
     dependencies {
-        classpath "net.ltgt.gradle:gradle-apt-plugin:0.6"
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
 
-apply plugin: "net.ltgt.apt"
-
 dependencies {
-    compile 'com.google.code.gson:gson:2.6.2'
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    compile "com.google.code.gson:gson:$gson_version"
 
-    compile 'net.lachlanmckee:gsonpath:2.3.0'
-    apt 'net.lachlanmckee:gsonpath-compiler:2.3.0'
-    apt project(":gsonpath-extension-android")
+    compile "net.lachlanmckee:gsonpath:$gsonpath_version"
+    kapt "net.lachlanmckee:gsonpath-compiler:$gsonpath_version"
+    compile project(":gsonpath-extensions")
+    kapt project(":gsonpath-extensions-compiler")
 
-    testCompile 'junit:junit:4.12'
+    testCompile "junit:junit:$junit_version"
 }
