+++ integration-test/src/main/java/gsonpath/extension/range/floatrange/FloatRangeModel.java
@@ -1,6 +1,5 @@
 package gsonpath.extension.range.floatrange;
 
-import android.support.annotation.FloatRange;
 import gsonpath.AutoGsonAdapter;
 
 interface FloatRangeModel {
@@ -18,13 +17,13 @@ interface FloatRangeModel {
     interface Inclusive {
         @AutoGsonAdapter
         interface FloatModel extends BaseFloatModel {
-            @FloatRange(from = 0, to = 5)
+            @android.support.annotation.FloatRange(from = 0, to = 5)
             Float getValue();
         }
 
         @AutoGsonAdapter
         interface DoubleModel extends BaseDoubleModel {
-            @FloatRange(from = 0, to = 5)
+            @gsonpath.extension.annotation.FloatRange(from = 0, to = 5)
             Double getValue();
         }
     }
@@ -32,13 +31,13 @@ interface FloatRangeModel {
     interface Exclusive {
         @AutoGsonAdapter
         interface FloatModel extends BaseFloatModel {
-            @FloatRange(from = 0, to = 5, fromInclusive = false, toInclusive = false)
+            @android.support.annotation.FloatRange(from = 0, to = 5, fromInclusive = false, toInclusive = false)
             Float getValue();
         }
 
         @AutoGsonAdapter
         interface DoubleModel extends BaseDoubleModel {
-            @FloatRange(from = 0, to = 5, fromInclusive = false, toInclusive = false)
+            @gsonpath.extension.annotation.FloatRange(from = 0, to = 5, fromInclusive = false, toInclusive = false)
             Double getValue();
         }
     }
