+++ integration-test/src/main/java/gsonpath/extension/range/intrange/IntRangeModel.java
@@ -1,6 +1,5 @@
 package gsonpath.extension.range.intrange;
 
-import android.support.annotation.IntRange;
 import gsonpath.AutoGsonAdapter;
 
 interface IntRangeModel {
@@ -11,13 +10,13 @@ interface IntRangeModel {
 
     @AutoGsonAdapter
     interface IntModel extends BaseModel<Integer> {
-        @IntRange(from = 0, to = 5)
+        @android.support.annotation.IntRange(from = 0, to = 5)
         Integer getValue();
     }
 
     @AutoGsonAdapter
     interface LongModel extends BaseModel<Long> {
-        @IntRange(from = 0, to = 5)
+        @gsonpath.extension.annotation.IntRange(from = 0, to = 5)
         Long getValue();
     }
 }
