+++ integration-test/src/main/java/gsonpath/extension/size/SizeModel.java
@@ -1,6 +1,5 @@
 package gsonpath.extension.size;
 
-import android.support.annotation.Size;
 import gsonpath.AutoGsonAdapter;
 
 import java.util.List;
@@ -20,13 +19,13 @@ interface SizeModel {
     interface MinAndMax {
         @AutoGsonAdapter
         interface ArrayModel extends BaseArrayModel {
-            @Size(min = 1, max = 3)
+            @gsonpath.extension.annotation.Size(min = 1, max = 3)
             Integer[] getValue();
         }
 
         @AutoGsonAdapter
         interface CollectionModel extends BaseCollectionModel {
-            @Size(min = 1, max = 3)
+            @android.support.annotation.Size(min = 1, max = 3)
             List<Integer> getValue();
         }
     }
@@ -34,13 +33,13 @@ interface SizeModel {
     interface Multiple {
         @AutoGsonAdapter
         interface ArrayModel extends BaseArrayModel {
-            @Size(multiple = 2)
+            @gsonpath.extension.annotation.Size(multiple = 2)
             Integer[] getValue();
         }
 
         @AutoGsonAdapter
         interface CollectionModel extends BaseCollectionModel {
-            @Size(multiple = 2)
+            @android.support.annotation.Size(multiple = 2)
             List<Integer> getValue();
         }
     }
@@ -48,13 +47,13 @@ interface SizeModel {
     interface ExactSize {
         @AutoGsonAdapter
         interface ArrayModel extends BaseArrayModel {
-            @Size(value = 1)
+            @gsonpath.extension.annotation.Size(value = 1)
             Integer[] getValue();
         }
 
         @AutoGsonAdapter
         interface CollectionModel extends BaseCollectionModel {
-            @Size(value = 1)
+            @android.support.annotation.Size(value = 1)
             List<Integer> getValue();
         }
     }
